backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain==0.1.0
langchain-google-genai==0.0.6
pandas==2.1.4
openpyxl==3.1.2
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0
aiofiles==23.2.1
